update birthday state
update birthday state

Fix: Prevent updating birthday when navigating from MyPage

This commit addresses an issue where the user's birthday was being
updated in the `AboutMeBirthdayScreen` even when the navigation
originated from the `MyPageScreen`.

The `isFromMyPage` parameter has been added to the navigation route and
ViewModel to control this behavior. Now, the birthday is only updated if
`isFromMyPage` is false.